+++ dst/app/routes/reservations.js
@@ -83,10 +83,10 @@ reservationsRouter.get('', permitScopes_1.default(['admin', 'reservations', 'res
             limit: (req.query.limit !== undefined) ? Math.min(req.query.limit, 100) : 100, page: (req.query.page !== undefined) ? Math.max(req.query.page, 1) : 1, sort: (req.query.sort !== undefined && req.query.sort.modifiedTime !== undefined)
                 ? { modifiedTime: req.query.sort.modifiedTime }
                 : undefined });
-        const totalCount = yield reservationRepo.count(searchCoinditions);
+        // const totalCount = await reservationRepo.count(searchCoinditions);
         const reservations = yield reservationRepo.search(searchCoinditions);
-        res.set('X-Total-Count', totalCount.toString())
-            .json(reservations);
+        // res.set('X-Total-Count', totalCount.toString())
+        res.json(reservations);
     }
     catch (error) {
         next(error);