+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- 予約検索結果のX-Total-Countを削除
+
 ### Deprecated
 
 ### Removed